feat(execution): A2+A3 — extract _session_outcome.py, finalize session.py facade
A2: Move _compute_success and _compute_outcome into _session_outcome.py.
Update _headless_result.py to import both functions directly from
_session_outcome rather than via the session facade (per issue spec).

A3: Rewrite session.py as a ~44-line pure re-export hub. All 16 symbols
remain importable from autoskillit.execution.session; no public API changes.
Removes all definitions — session.py now owns zero function/class/constant
definitions.
